+++ src/test/java/uk/gov/pay/connector/queue/tasks/handlers/AdyenCancellationNotificationHandlerTest.java
@@ -16,6 +16,7 @@ import uk.gov.pay.connector.charge.model.domain.Charge;
 import uk.gov.pay.connector.charge.model.domain.ChargeStatus;
 import uk.gov.pay.connector.gateway.adyen.webhook.AdyenNotificationService;
 import uk.gov.pay.connector.gateway.processor.ChargeNotificationProcessor;
+import uk.gov.pay.connector.queue.tasks.handlers.adyen.AdyenCancellationNotificationHandler;
 
 import java.time.Instant;
 import java.time.ZoneId;
